[parser] New utils directory New utils namespace Moved Container and Wrapper classes to utils Defined namesapces
