+++ Makefile
@@ -5,6 +5,7 @@ TEST_NAME	=	$(NAME)_test
 ### WEBSERV SOURCES ###
 #######################
 HDRS_DIR		=	include
+HDRS_UTILS_DIR	=	$(HDRS_DIR)/utils
 HDRS_CONFIG_DIR	=	$(HDRS_DIR)/configurations
 
 SRCS_DIR		=	srcs
@@ -15,10 +16,10 @@ MAIN_SRC	=	$(SRCS_DIR)/main.cpp
 MAIN_OBJ	=	$(MAIN_SRC:.cpp=.o)
 MAIN_DEPEN	=	$(MAIN_SRC:.cpp=.d)
 
-SRCS		=	$(addprefix $(SRCS_UTILS_DIR)/, utils_1.cpp)\
-					$(addprefix $(SRCS_CONFIG_DIR)/, configuration.cpp configuration_host.cpp exceptions.cpp parser.cpp route_configuration.cpp server_configuration.cpp utils.cpp)
-HDRS		=	$(addprefix $(HDRS_DIR)/, utils.hpp)\
-					$(addprefix $(HDRS_CONFIG_DIR)/, configuration_host.hpp	configuration.hpp container.hpp exceptions.hpp parser.hpp route_configuration.hpp server_configuration.hpp utils.hpp wrapper.hpp)
+SRCS		=	$(addprefix $(SRCS_UTILS_DIR)/, container.cpp exceptions.cpp utils.cpp wrapper.cpp)\
+					$(addprefix $(SRCS_CONFIG_DIR)/, configuration.cpp configuration_host.cpp parser.cpp route_configuration.cpp server_configuration.cpp utils.cpp)
+HDRS		=	$(addprefix $(HDRS_UTILS_DIR)/, container.hpp exceptions.hpp utils.hpp wrapper.hpp)\
+					$(addprefix $(HDRS_CONFIG_DIR)/, configuration_host.hpp	configuration.hpp parser.hpp route_configuration.hpp server_configuration.hpp utils.hpp)
 OBJS		=	$(SRCS:.cpp=.o)
 DEPEN		=	$(SRCS:.cpp=.d)
