+++ include/configurations/parser.hpp
@@ -3,7 +3,7 @@
 #include "utils.hpp"
 #include "configuration.hpp"
 
-namespace wsrv{
+namespace MAIN_NAMESPACE{
 
 class Parser{
 public:
@@ -14,13 +14,13 @@ public:
     ~Parser();
 
     Parser& operator=(const Parser& other);
-    const configuration::Configuration& operator->();
+    const CONFIG_NAMESPACE::Configuration& operator->();
 
     static Parser parseFile(const std::string& inputFile);
-    const configuration::Configuration& getConfiguration() const;
+    const CONFIG_NAMESPACE::Configuration& getConfiguration() const;
 
 private:
-    configuration::Configuration configuration_;
+    CONFIG_NAMESPACE::Configuration configuration_;
 
 };
